adding feature

+++ apis/apis/api-hope/src/main/java/com/hope/api_hope/controller/UserController.java
@@ -21,7 +21,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<UserDTO> getUserById(@PathVariable int id){
+    public ResponseEntity<UserDTO> getUserById(@PathVariable Long id){
         UserDTO userDTO = service.getUserById(id);
 
         return userDTO != null ? ResponseEntity.ok(userDTO) : ResponseEntity.notFound().build();
@@ -33,14 +33,14 @@ public class UserController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<UserDTO> updateUser(@PathVariable int id,@RequestBody UserDTO userDTO){
+    public ResponseEntity<UserDTO> updateUser(@PathVariable Long id,@RequestBody UserDTO userDTO){
         UserDTO updateUser = service.updateUser(id, userDTO);
 
         return updateUser != null ? ResponseEntity.ok(updateUser) : ResponseEntity.notFound().build();
     }
 
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable int id){
+    public void deleteUser(@PathVariable Long id){
         service.deleteUser(id);
     }
 }
